source/data_engineering/paper_scraper/requirements.txt
bs4
kafka-python-ng
lxml
aiohttp[speedups]
aiokafka